// src/app/interfaces/register-doctor.ts
export interface RegisterDoctor {
    username: string;
    password: string;
    confirmPassword: string;
    email: string;
    firstname: string;
    lastname: string;
    phoneNumber: string;
    city: string;
    address: string;
    streetNumber: number;
    speciality: string;
    appointmentDuration: number;  
}